+++ app/build.gradle
@@ -42,10 +42,10 @@ dependencies {
 
     implementation 'androidx.core:core-ktx:1.12.0'
     implementation 'androidx.appcompat:appcompat:1.6.1'
-    implementation 'com.google.android.material:material:1.10.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     implementation 'androidx.preference:preference:1.2.1'
     implementation 'androidx.recyclerview:recyclerview-selection:1.1.0'
+    implementation 'androidx.navigation:navigation-fragment-ktx:2.5.2'
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
@@ -55,4 +55,7 @@ dependencies {
     implementation  'com.squareup.retrofit2:converter-gson:2.9.0'
     implementation 'androidx.core:core-ktx:1.12.0'
     implementation 'io.insert-koin:koin-android:3.5.0'
+    implementation "androidx.fragment:fragment-ktx:1.5.7"
+    implementation "androidx.navigation:navigation-ui-ktx:2.5.3"
+    implementation 'com.google.android.material:material:1.8.0'
 }
