+++ app/src/main/AndroidManifest.xml
@@ -30,20 +30,6 @@
             android:name=".player.ui.PlayerActivity"
             android:exported="false">
         </activity>
-        <activity
-            android:name=".media.ui.MediaActivity"
-            android:exported="false">
-        </activity>
-        <activity
-            android:name=".search.ui.SearchActivity"
-            android:exported="false">
-        </activity>
-        <activity
-            android:name=".setting.ui.SettingsActivity"
-            android:exported="false"
-            android:label="@string/title_activity_settings" >
-        </activity>
-
     </application>
 
 </manifest>
