--- app/src/main/java/com/example/playlistmaker/media/ui/MediaActivity.kt
@@ -1,41 +1,0 @@
-package com.example.playlistmaker.media.ui
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import com.example.playlistmaker.R
-import com.example.playlistmaker.databinding.ActivityMediaBinding
-import com.google.android.material.tabs.TabLayoutMediator
-
-class MediaActivity : AppCompatActivity() {
-
-    private var _binding: ActivityMediaBinding? = null
-    private val binding get() = _binding!!
-    private var _tabMediator: TabLayoutMediator? = null
-    private val tabMediator get() = _tabMediator!!
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        _binding = ActivityMediaBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        binding.viewPager.adapter = MediaViewPagerAdapter(supportFragmentManager, lifecycle)
-
-        binding.btBack.setOnClickListener {
-            finish()
-        }
-
-        _tabMediator = TabLayoutMediator(binding.tabLayout, binding.viewPager) { tab, position ->
-            when (position) {
-                0 -> tab.setText( R.string.fav_tracks)
-                1 -> tab.setText(R.string.playlists)
-            }
-        }
-
-        tabMediator.attach()
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        _tabMediator?.detach()
-    }
-}
